Rename SystemAudioDevicesChangeEventTargetComposite
Issue #228
